+++ goarray/goarray.go
@@ -16,3 +16,16 @@ func TurnString(ss []string) (s string) {
 	}
 	return s
 }
+
+// RemoveDuplicateElement 去重
+func RemoveDuplicateElement(ss []string) []string {
+	result := make([]string, 0, len(ss))
+	temp := map[string]struct{}{}
+	for _, item := range ss {
+		if _, ok := temp[item]; !ok {
+			temp[item] = struct{}{}
+			result = append(result, item)
+		}
+	}
+	return result
+}